+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [1.0.0] - 2021-05-01
 - [BREAKING] New widgets take a `&'static str` for a label when initialized, to use dynamic labels, use set_label or with_label. This reflects the FLTK api.
+- Add a Column and Row widgets which support auto_layout by default, but require that widgets be added using add().
 
 
 ## [0.16.5] - 2021-04-10